[codex] Clean up Rodinia GPU benchmark correctness (#5)
## Summary

This PR keeps the Rodinia modernization focused on real benchmark
execution paths and removes the simulator/PTX compatibility work from
the local cleanup branch. It preserves the correctness fixes found
during CUDA, OpenCL, and OpenMP validation, adds a CUDA run script
matching the existing OpenMP/OpenCL runners, and keeps generated
benchmark output out of the source tree.

Main changes:

- Add `scripts/run_cuda.sh` and update the OpenMP/OpenCL runners so
benchmark outputs go under the selected build output directory instead
of polluting the source tree.
- Remove tracked generated result files and ignore `build-*`
directories.
- Remove stale OpenCL PTX dump hooks and PTX-oriented README text from
BFS, CFD, hybridsort, and streamcluster.
- Keep real correctness fixes found during cross-variant validation,
including the CUDA BFS input source handling and the OpenMP precomputed
double CFD missing `.z` field load.
- Keep warning and robustness cleanups in CUDA/OpenCL/OpenMP benchmark
code and bundled legacy helper code where they were needed for current
toolchains.
- Update the comparison helper for current output locations and for
streamcluster's stochastic center output format.

## Why

The prior branch had drifted into simulator/compatibility-oriented
changes. That is not the direction we want for this suite: the benchmark
should prioritize real GPU/OpenMP/OpenCL execution and comparable
performance behavior on current toolchains. This PR keeps that focus
while retaining the concrete correctness and cleanup fixes that were
independently validated on a real NVIDIA GPU node.

## Correctness Notes

- `streamcluster` is a stochastic local-search benchmark. The exact
chosen centers can vary across implementations while still representing
a real run. The comparison now checks structural validity: expected
center-count range, total center weight, and coordinate dimensionality.
In the final rerun, OpenCL and CUDA matched exactly for the generated
center set.
- `myocyte` shows small CUDA numerical drift relative to the CPU/OpenCL
variants in the ODE state. The final comparison accepts this with an
absolute bound of `0.2`; the observed maximum absolute difference was
`0.140671`.
- CUDA `kmeans` and CUDA `nw` do not currently emit the same result
artifact format as the other variants, so the semantic comparison only
includes variants with comparable files for those checks.

## Validation

Validated on this node with:

- NVIDIA GPU: `Quadro M4000`, driver `580.159.03`, 8192 MiB
- CUDA toolkit: `nvcc` 12.3 (`V12.3.103`)
- OpenCL platforms/devices visible through `clinfo`: NVIDIA CUDA `Quadro
M4000`, Intel OpenCL Graphics, and POCL CPU

Builds:

```sh
cmake --build build-gcc -j32
cmake --build build-cuda -j32
cmake --build build-omp-clang-cuda -j32
```

Full runs:

```sh
scripts/run_openmp.sh --build-dir build-gcc --output-dir build-gcc/run_outputs/pr_final_omp_gcc --threads 32
scripts/run_opencl.sh --build-dir build-gcc --output-dir build-gcc/run_outputs/pr_final_opencl
scripts/run_cuda.sh --build-dir build-cuda --output-dir build-cuda/run_outputs/pr_final_cuda --device 0
OMP_TARGET_OFFLOAD=MANDATORY scripts/run_openmp.sh --build-dir build-omp-clang-cuda --output-dir build-omp-clang-cuda/run_outputs/pr_final_omp_offload --threads 32 --skip-mummergpu
```

Targeted reruns after final fixes:

```sh
cmake --build build-gcc --target openmp_cfd_pre_euler3d_cpu_double -j32
cmake --build build-omp-clang-cuda --target openmp_cfd_pre_euler3d_cpu_double -j32
scripts/run_openmp.sh --build-dir build-gcc --output-dir build-gcc/run_outputs/pr_final_omp_gcc --threads 32 --only cfd_pre_double
OMP_TARGET_OFFLOAD=MANDATORY scripts/run_openmp.sh --build-dir build-omp-clang-cuda --output-dir build-omp-clang-cuda/run_outputs/pr_final_omp_offload --threads 32 --skip-mummergpu --only cfd_pre_double
scripts/run_opencl.sh --build-dir build-gcc --output-dir build-gcc/run_outputs/pr_final_opencl --only bfs --only cfd --only gaussian --only hybridsort --only nn --only streamcluster
```

Semantic cross-variant comparison covered 29 checks across OpenMP,
OpenMP offload, OpenCL, and CUDA result artifacts where output formats
are comparable. The final comparison completed with:

```text
validated 29 checks
```

Hygiene checks:

```sh
git diff --check
git diff --cached --check
bash -n scripts/run_openmp.sh scripts/run_opencl.sh scripts/run_cuda.sh scripts/compare_openmp_opencl.sh
find cuda opencl openmp -type f \( -name 'result.txt' -o -name 'output.txt' -o -name 'image_out.pgm' -o -name 'hybridinput.txt' -o -name 'hybridoutput.txt' -o -name 'PD.txt' -o -name '*.timing' \) -print
rg -n "PTX_MSG|cl\.ptx|generate ptx|print the ptx" opencl scripts CMakeLists.txt
```

The artifact and PTX searches returned no source-tree matches.

+++ openmp/backprop/imagenet.c
@@ -10,6 +10,6 @@ void load(BPNN *net) {
   int nr = layer_size;
 
   for (int i = 0, k = 1; i < nr; i++, k++) {
-    units[k] = (float)rand() / (float)RAND_MAX;
+    units[k] = (float)((double)rand() / (double)RAND_MAX);
   }
 }
